+++ src/components/App.js
@@ -1,8 +1,8 @@
 import { useSelector } from "react-redux"
 
-import MenuScreen from "./components/MenuScreen"
-import GameScreen from "./components/GameScreen"
-import VictoryModal from "./components/VictoryModal"
+import MenuScreen from "./MenuScreen"
+import GameScreen from "./GameScreen"
+import VictoryModal from "./VictoryModal"
 
 export default function App() {
     const isStarted = useSelector(s => s.startTime !== undefined)
